Use conditional imports for web and dart:io (stub) implementations
This is so that unit tests can run without importing dart:js_interop and related packages.
