+++ web/_packages/web_startup_analyzer/example/pubspec.yaml
@@ -3,7 +3,8 @@ description: "flutter_web_startup_analyzer example"
 publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 environment:
-  sdk: '>=3.4.0-16.0.dev <4.0.0'
+  sdk: ^3.2.0
+  flutter: ^3.16.0
 dependencies:
   flutter:
     sdk: flutter
